feat(PlanCalculator): move components to new files
//feat(PlanCalculator): move components to new files BIS
