+++ src/components/ChangePlan/PlanCalculator/PlanCalculator.js
@@ -1,227 +1,21 @@
 import React, { useReducer, useEffect, useState } from 'react';
-import { Slider } from '../shared/Slider/Slider';
-import { TooltipContainer } from '../TooltipContainer/TooltipContainer';
-import { InjectAppServices } from '../../services/pure-di';
-import { Loading } from '../Loading/Loading';
-import { FormattedMessage, useIntl } from 'react-intl';
+import { Slider } from '../../shared/Slider/Slider';
+import { TooltipContainer } from '../../TooltipContainer/TooltipContainer';
+import { InjectAppServices } from '../../../services/pure-di';
+import { Loading } from '../../Loading/Loading';
+import { useIntl } from 'react-intl';
 import queryString from 'query-string';
-import { extractParameter, getPlanFee, thousandSeparatorNumber } from '../../utils';
+import { extractParameter } from '../../../utils';
 import { useRouteMatch, Link } from 'react-router-dom';
-import useTimeout from '../../hooks/useTimeout';
+import useTimeout from '../../../hooks/useTimeout';
 import * as S from './PlanCalculator.styles';
-
-const NavigatorTabs = ({ tabs, pathType, selectedPlanType }) => {
-  const intl = useIntl();
-  const _ = (id, values) => intl.formatMessage({ id: id }, values);
-
-  const getTypePlanDescriptionWithTooltip = (type) => {
-    switch (type) {
-      case 'prepaid':
-      case 'subscribers':
-      case 'monthly-deliveries':
-        return (
-          <>
-            {_(`plan_calculator.plan_type_${type.replace('-', '_')}`)}{' '}
-            <div className="dp-tooltip-container">
-              <span className="ms-icon icon-info-icon"></span>
-              <div className="dp-tooltip-top">
-                <span>{_(`plan_calculator.plan_type_${type.replace('-', '_')}_tooltip`)}</span>
-              </div>
-            </div>
-          </>
-        );
-      default:
-        return '';
-    }
-  };
-
-  return (
-    <nav className="tabs-wrapper">
-      <ul className="tabs-nav">
-        {tabs.map((type, index) => (
-          <li className="tab--item" key={index}>
-            <Link
-              to={`/plan-selection/${pathType}/${type}`}
-              className={type === selectedPlanType ? 'tab--link active' : 'tab--link'}
-            >
-              {getTypePlanDescriptionWithTooltip(type)}
-            </Link>
-          </li>
-        ))}
-      </ul>
-    </nav>
-  );
-};
-
-const Discounts = ({ discountsList, handleChange }) => {
-  const [selectedDiscount, setSelectedDiscount] = useState(discountsList[0]);
-  const intl = useIntl();
-  const _ = (id, values) => intl.formatMessage({ id: id }, values);
-
-  const getDiscountDescription = (discountDescription) => {
-    switch (discountDescription) {
-      case 'monthly':
-      case 'quarterly':
-      case 'half-yearly':
-      case 'yearly':
-        return _('plan_calculator.discount_' + discountDescription.replace('-', '_'));
-      default:
-        return '';
-    }
-  };
-
-  return (
-    <>
-      <div className="dp-wrap-subscription">
-        <h4>{_('plan_calculator.discount_title')}</h4>
-        <ul>
-          {discountsList.map((discount, index) => (
-            <li key={index}>
-              <button
-                key={index}
-                className={`dp-button button-medium ${
-                  discount.id === selectedDiscount.id ? 'btn-active' : ''
-                }`}
-                onClick={() => {
-                  handleChange(discount);
-                  setSelectedDiscount(discount);
-                }}
-              >
-                {getDiscountDescription(discount.description)}
-              </button>
-              {discount.discountPercentage ? (
-                <span className="dp-discount">{`${discount.discountPercentage}% OFF`}</span>
-              ) : (
-                <></>
-              )}
-            </li>
-          ))}
-        </ul>
-      </div>
-    </>
-  );
-};
-
-const BannerUpgrade = ({ currentPlan, currentPlanList }) => {
-  const getBannerInfo = (type) => {
-    const bannerInfo = { messageId: `plan_calculator.banner_for_${type.replace('-', '_')}` };
-    switch (type) {
-      case 'prepaid':
-      case 'subscribers':
-        return {
-          ...bannerInfo,
-          link: `/plan-selection/${currentPlan.featureSet}/monthly-deliveries`,
-        };
-      case 'monthly-deliveries':
-        // TODO: define where to go in this case
-        return { ...bannerInfo, link: `/email-marketing-agencies` };
-      default:
-        return `plan_calculator.banner_for_unknown`;
-    }
-  };
-
-  if (currentPlan.id === currentPlanList[currentPlanList.length - 1].id) {
-    const bannerInfo = getBannerInfo(currentPlan.type);
-    return (
-      <div className="dp-calc-message">
-        <p>
-          <FormattedMessage
-            id={bannerInfo.messageId}
-            values={{
-              Link: (chunk) => (
-                <Link to={bannerInfo.link}>
-                  <strong>{chunk}</strong>
-                </Link>
-              ),
-            }}
-          />
-        </p>
-      </div>
-    );
-  }
-  return <></>;
-};
-
-const PlanPriceWithoutDiscounts = ({ planData }) => {
-  const intl = useIntl();
-  const formatedFee = thousandSeparatorNumber(intl.defaultLocale, getPlanFee(planData.plan));
-  return (
-    <>
-      {planData.discount?.discountPercentage ? (
-        <span className="dp-price-old">
-          <span className="dp-price-old-money">US$</span>
-          <span className="dp-price-old-amount">{formatedFee}</span>
-        </span>
-      ) : (
-        <></>
-      )}
-    </>
-  );
-};
-
-const PlanPricePerMonth = ({ planData }) => {
-  const intl = useIntl();
-  const _ = (id, values) => intl.formatMessage({ id: id }, values);
-  const discountPercentage = planData.discount?.discountPercentage;
-  const planFee = getPlanFee(planData.plan);
-  const planFeeWithDiscount = thousandSeparatorNumber(
-    intl.defaultLocale,
-    discountPercentage ? planFee * (1 - discountPercentage / 100) : planFee,
-  );
-
-  return (
-    <>
-      <h2 className="dp-price-large">
-        <span className="dp-price-large-money">US$</span>
-        <span className="dp-price-large-amount">{planFeeWithDiscount}</span>
-      </h2>
-      <span className="dp-for-time">{_('plan_calculator.per_month')}</span>
-    </>
-  );
-};
-
-const PlanAgreement = ({ planData }) => {
-  const intl = useIntl();
-  const _ = (id, values) => intl.formatMessage({ id: id }, values);
-  const discountPercentage = planData.discount?.discountPercentage;
-  const monthsAmmount = planData?.discount?.monthsAmmount;
-  const planFee = Math.round(getPlanFee(planData.plan));
-  const planFeeWithDiscount = thousandSeparatorNumber(
-    intl.defaultLocale,
-    planFee * (1 - discountPercentage / 100) * monthsAmmount,
-  );
-
-  const getAgreementDescription = (discountDescription) => {
-    switch (discountDescription) {
-      case 'quarterly':
-      case 'half-yearly':
-      case 'yearly':
-        return _('plan_calculator.with_' + discountDescription.replace('-', '_') + '_discount');
-      default:
-        return '';
-    }
-  };
-
-  return (
-    <div className="dp-agreement">
-      {discountPercentage ? (
-        <p>
-          {getAgreementDescription(planData.discount.description)}
-          <strong>
-            {' '}
-            US$
-            {planFeeWithDiscount}
-          </strong>
-        </p>
-      ) : null}
-      <p>{_('plan_calculator.discount_clarification')}</p>
-    </div>
-  );
-};
-
-const PlanPrice = ({ children }) => {
-  return <div className="dp-price--wrapper">{children}</div>;
-};
+import { NavigatorTabs } from './NavigatorTabs/NavigatorTabs';
+import { Discounts } from './Discounts/Discounts';
+import { BannerUpgrade } from './BannerUpgrade/BannerUpgrade';
+import { PlanPriceWithoutDiscounts } from './PlanPriceWithoutDiscounts/PlanPriceWithoutDiscounts';
+import { PlanPricePerMonth } from './PlanPricePerMonth/PlanPricePerMonth';
+import { PlanAgreement } from './PlanAgreement/PlanAgreement';
+import { PlanPrice } from './PlanPrice/PlanPrice';
 
 const PlanCalculator = ({ location, dependencies: { planService, appSessionRef } }) => {
   const safePromoId = extractParameter(location, queryString.parse, 'promo-code') || '';
