+++ src/components/ChangePlan/PlanCalculator/PlanPrice/PlanPrice.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const PlanPrice = ({ children }) => {
+  return <div className="dp-price--wrapper">{children}</div>;
+};
